try to isolate all the csv_string_model stuff
