+++ lib/csv_row_model/model/base.rb
@@ -34,19 +34,6 @@ module CsvRowModel
       def abort?
         false
       end
-
-      class_methods do
-        # @return [Class] the Class with validations of the csv_string_model
-        def csv_string_model_class
-          @csv_string_model_class ||= inherited_custom_class(:csv_string_model_class, CsvStringModel)
-        end
-
-        protected
-        # Called to add validations to the csv_string_model_class
-        def csv_string_model(&block)
-          csv_string_model_class.class_eval(&block)
-        end
-      end
     end
   end
 end
