--- lib/csv_row_model/model/csv_string_model.rb
@@ -1,7 +1,0 @@
-module CsvRowModel
-  module Model
-    class CsvStringModel < OpenStruct
-      include ActiveWarnings
-    end
-  end
-end
